refactor: salon/hooks 수정(DURI-85)
