+++ packages/utils/src/apis/duri/auth.ts
@@ -3,6 +3,6 @@ import { authInstance, NaverLoginResponse } from "@duri-fe/utils";
 /** 네이버 소셜 로그인 */
 export const duriNaverLogin = async (): Promise<NaverLoginResponse> => {
   const response = await authInstance.get('authorization/naver-user');
-  console.log(response);
+  window.location.href = response.data;
   return response.data;
 }
